+++ entrypoints/extpay.content.ts
@@ -1,7 +1,0 @@
-import "extpay";
-
-export default defineContentScript({
-  matches: ["https://extensionpay.com/*"],
-  runAt: "document_start",
-  main() {},
-});